+++ media/main.js
@@ -45,6 +45,7 @@ $(document).ready(function () {
       )
       .replaceAll("/@", "/");
   };
+
   const addPRsToDoc = (prs) => {
     $("#ghHolder").append(
       "<button class='run-watermelon'>Run Watermelon</button><br/>"
@@ -90,7 +91,7 @@ $(document).ready(function () {
         <div>
           <div class="pr-owner">
             <p class="pr-poster" title="View this user on github">
-              Author: <a href="${pr.userLink}">${pr.user}</a>
+              <a class="pr-author-combo" href="${pr.userLink}"><img class='pr-author-img' src="${pr.userImage}" />${pr.user}</a>
             </p>
             <p class="pr-date">
               ${new Date(pr.created_at).toLocaleDateString("en-us", {